calculate calories from database in showcalscreen

+++ lib/model/food.dart
@@ -1,6 +1,6 @@
 class Foods{
   String? name;
-  double? calories;
+  int? calories;
   int? amount;
   Foods({this.name,this.calories,this.amount});
 }
